refactor: updated modal service
