+++ src/app/modules/customer-vehicle/components/customer-vehicle-table/customer-vehicle-table.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CustomerVehicle } from '../../../../core/models/CustomerVehicle';
 import { CustomerVehicleService } from '../../../../core/services/customer-vehicle/CustomerVehicle.service';
-import { ModalService } from '../../../services/modal/modal.service';
+import { ModalService } from '../../../../shared/services/modal/modal.service';
 
 @Component({
   selector: 'app-customer-vehicle-table',
